Phase3/requirements.txt
kafka-python>=2.0.2
python-dotenv>=1.0.0
requests>=2.31.0

# Data & ML
pandas>=2.0.0
numpy>=1.24.0
torch>=2.1.0
joblib>=1.3.0

# Dashboard
streamlit>=1.29.0
plotly>=5.18.0

# LLM providers
openai>=1.7.0
anthropic>=0.8.0
google-generativeai>=0.5.0
